Implement compilation automatization buttons
This commit adds 5 new buttons: BSP, VIS, LIGHT, Launch after compile
and Compile selected.

The first three buttons replace the BSP, VIS and LIGHT labels and can be
toggled. Pressing the "Compile selected" button compiles the stages that
are enabled. If "Launch after compile" is checked, the map will be
launched once compilation is done.

Buttons are re-themed to allow for the "pressed" state of buttons which
didn't work as expected with the previous visuals.

"Compiling selected" can be run by pressing F4, clearing the console is
now done with F12.