Add licenses screen
Implemented a top bar for switching between the Quake compiler screen
and a screen with licenses. The licenses screen shows the application's
license, the Godot Engine license and the Godot third party licenses.

This commit also adds the export_presets.cfg, because building the
project correctly requires including the licensing text files, which are
not included in builds by default.

The default window height has been increased to 670 due to the new top
bar.