Move hotkeys from Main to Compiler and Runner
This commit moves the hotkeys for compiling and running from
Main to the Compiler and Runner scripts. Main is also deleted,
as there is no code left.

This commit also fixes a bug where pressing Alt+F4 would start compiling
the selected stages and then quit the application. The application now
checks if Alt is pressed together with F4 and quits right away.