Fix batch compiling bug on Linux
When batch compiling with q3map2 on Linux, the process detection for
q3map2 did not work properly and exited too early, causing a new queue
item to start running while q3map2 was still compiling the current queue
item. This seems to happen because q3map2 can spawn new subprocesses
with new PIDs.

This commit fixes that by looking for running processes with the
compiler's file name instead of relying on PIDs.

The fix is not 100% ideal, as there could in theory be other q3map2
processes running at the same time, which are compiling different maps.
Quakompiler would then wait for all q3map2 instances to finish before
continuing the batch operation. This is not critical and very unlikely
though, so it's fine for now.